Optimize Imports and restructure Code

+++ app/src/main/java/com/example/soudest/helper/location.java
@@ -1,21 +1,22 @@
 package com.example.soudest.helper;
 
-import com.google.android.gms.maps.model.LatLng;    //For get_cords_to_Location
+import com.google.android.gms.maps.model.LatLng;
+
 import java.util.Random;
 
 public final class location {
     static Random r = new Random();
 
-    private location(){
+    private location() {
         throw new IllegalStateException("No instances allowed!");
     }
 
     //Returns Fake Coordinates for a given Adress within Weingarten
-    public static LatLng getFakeCordsToLocation(String adress){
+    public static LatLng getFakeCordsToLocation(String adress) {
 
-        double north = 47.808+(double)((double)getRandomNumberInRange(0,1)/100);
-        double east = 9.639+(double)((double)getRandomNumberInRange(0,1)/100);
-        LatLng pos = new LatLng(north,east);
+        double north = 47.808 + ((double) getRandomNumberInRange(0, 1) / 100);
+        double east = 9.639 + ((double) getRandomNumberInRange(0, 1) / 100);
+        LatLng pos = new LatLng(north, east);
         return pos;
     }
